move generation to a better spot i think
