+++ src/lib.rs
@@ -1,4 +1,3 @@
-mod generation;
 mod models;
 mod utils;
