+++ src/models/raw/generation.rs
