+++ src/models/raw/mod.rs
@@ -1,3 +1,4 @@
+pub mod generation;
 pub mod models;
 mod quantized_nn;
 mod quantized_var_builder;
